+++ q1pulse/__init__.py
@@ -1,5 +1,5 @@
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 
 from .instrument import Q1Instrument, set_exception_on_overload
 from .lang.math_expressions import Lsr as lsr
-
+from .util.dumper import q1asm_dump